requirements.txt
Django==1.11.4
django-debug-toolbar==1.8
django-extensions==1.6.7
django-phonenumber-field==1.3.0
psycopg2==2.7.1
djangorestframework==3.6.4
markdown==2.6.9
django-filter==1.0.4
Pillow==4.2.1
pytz==2017.2
ipython==5.4.1
ipdb==0.10.3
djangocms-admin-style==1.2.7
django-registration-redux==1.7
django-localflavor==1.5.3
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
django-cors-headers==2.1.0
django-sendgrid-v5==0.6.84